[libs] E: Move Binding for getpeername()
This commits moves the binding for the getpeername() system calls to the
syscall crate
