+++ src/libs/syscall/src/sys/socket/bindings/mod.rs
@@ -8,6 +8,7 @@
 pub mod accept;
 pub mod bind;
 pub mod connect;
+pub mod getpeername;
 pub mod listen;
 pub mod setsockopt;
 pub mod socket;
